farseer/backend/requirements.txt
fastapi
uvicorn
python-dotenv
ell-ai
openai